requirements.txt
flask
flask-cors
requests
websockets
python-dotenv
yfinance
beautifulsoup4
lxml
pandas
